third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/the-page/body-margin-3b-expected.txt

This is a testharness.js-based test.
[FAIL] Test that body's rightmargin and bottommargin attributes are ignored in quirks mode
  assert_equals: expected "8px" but got "100px"
Harness: the test ran to completion.
